pyproject.toml
[project]
name = "claude-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alpaca-py>=0.43.2",
    "google-genai>=1.71.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-telegram-bot>=22.7",
    "structlog>=25.5.0",
]

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
]
